--- src/API/Auth/signUp.ts
@@ -1,4 +1,0 @@
-
-export default function signUp(){
-  
-}
